refactor(graph): split CsrIndex into submodule and widen edge label ids to u32
Split nodedb-graph/src/csr/index.rs into a module directory (types,
interning, mutation, lookup, tests) so each concern fits within the
500-line file limit and can be reviewed independently.

Widen edge label ids from u16 to u32 throughout the csr layer — labels
intern, compaction, persist, statistics, weights, slice_accessors,
traversal, and the Data Plane algo and pattern consumers. The old u16
ceiling (65 535 distinct labels) is a silent wrap; the new ceiling is
~4 billion, and the typed GraphError::LabelOverflow makes overflow loud.

Add nodedb-graph/src/error.rs with GraphError and MAX_EDGE_LABELS.
Add thiserror dependency to nodedb-graph.

All add_edge / add_edge_weighted call sites updated to propagate
Result<(), GraphError> instead of ignoring the return value.

+++ nodedb-graph/src/csr/compaction.rs
@@ -10,8 +10,8 @@ impl CsrIndex {
     /// The old arrays are dropped, freeing memory. O(E) where E = total edges.
     pub fn compact(&mut self) {
         let n = self.id_to_node.len();
-        let mut new_out_edges: Vec<Vec<(u16, u32)>> = vec![Vec::new(); n];
-        let mut new_in_edges: Vec<Vec<(u16, u32)>> = vec![Vec::new(); n];
+        let mut new_out_edges: Vec<Vec<(u32, u32)>> = vec![Vec::new(); n];
+        let mut new_in_edges: Vec<Vec<(u32, u32)>> = vec![Vec::new(); n];
         let mut new_out_weights: Vec<Vec<f64>> = if self.has_weights {
             vec![Vec::new(); n]
         } else {
